+++ features/config/pull_branch_strategy/update.feature
@@ -1,7 +1,7 @@
 Feature: configure the pull_branch_strategy
 
   Scenario Outline:
-    When I run "git-town pull-branch-strategy <VALUE>"
+    When I run "git-town config pull-branch-strategy <VALUE>"
     Then setting "pull-branch-strategy" is now "<VALUE>"
 
     Examples:
